Move files to the new 'main' package
