+++ app/src/main/java/ecb/manifest/kowalski/main/ui/navigation/DrawerNavigationItem.kt
@@ -1,4 +1,4 @@
-package ecb.manifest.kowalski.ui.navigation
+package ecb.manifest.kowalski.main.ui.navigation
 
 import androidx.compose.ui.graphics.vector.ImageVector
